feat: two-step POS checkout with phone-first customer entry
- Two-step Cart -> Payment wizard; total + primary action pinned in a footer.
- Customer entry: two always-on fields (phone, name). A valid phone runs a
  debounced exact lookup that prefills and locks the name for returning
  customers; otherwise the cashier types a new name.
- POSTOrderSchema sends customer{name,phone} instead of customer_id. The server
  find-or-creates the Customer by phone inside the Order transaction (atomic,
  no orphans) and Title Cases names at the insert boundary. See ADR-0011.
- Drop CustomerAutocomplete and the select-based customer picker.
